Project Feedback / Issue
Postalicious is the only project for the matrix section for *Backend* under *Priority 1*: "Understand API documentation and write code that consumes it", under that skill set is:
```
1. [ ] Send basic authentication headers in an HTTP request
2. [ ] Use OAuth to authenticate against an API like Github's
3. [X] Use an API to read data from a remote service
4. [X] Use an API to create a resource on a remote service
5. [ ] Make API requests using cURL and Postman
6. [ ] Familar with SOAP and REST
```
I don't believe this project, Postalicious, covered all of these.
#1: We learned how to send headers, but the description in the matrix seems to imply something special about a header with the key "authentication"?
#2: OAuth wasn't used.
#3 & #4: I *think* maybe we did this. Is this want the sandbox server and postalicious app are doing? I don't see this written explicitly anywhere. I'm not sure if this is what I just learned or not.
#5: Didn't learn what cURL is or how to use it. Did learn how to use postman however.
#6: Never saw SOAP anywhere, I *think* that the HTTP verbs like "GET, POST, PATCH" etc are "REST" but I'm not 100% sure. I didn't see this correlation explicitly listed anywhere.

Also in *Backend* under *Priority 1* this project is listed under the description: "Basic codes, headers, post/get, URL, and parsing" with the following skillset:
```
1. [X] anatomy of a URL
2. [X] verbs: get, post
3. [X] headers: Accepts, Content-type
4. [X] status codes: 200, 404, 500
5. [X] params
6 [X] post body parsing
7. [X] articulate the difference between a standard HTTP request and an AJAX / XHR request
8. [X] list the most commonly used HTTP verbs
9. [ ] list the most commonly used HTTP headers
10. [ ] explain at a high level how Authentication via HTTP Cookies works
```
Most of these stats are wonderfully covered by this project. only 9. and 10. I don't feel like **I** learned, but maybe I just missed that. the Accept header was the only one covered by this project. As far as authentication and cookies, not covered at all.